Fix PillSheetModifiedHistoryCreatePillSheetAction
