--- lib/domain/calendar/components/pill_sheet_modified_history/components/pill_sheet_modified_history_created_pill_sheet_action.dart
@@ -1,55 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:pilll/components/atoms/font.dart';
-import 'package:pilll/components/atoms/text_color.dart';
-import 'package:pilll/domain/calendar/components/pill_sheet_modified_history/components/pill_sheet_modified_history_date_component.dart';
-import 'package:pilll/entity/pill_sheet_modified_history_value.dart';
-
-class PillSheetModifiedHistoryCreatePillSheetAction extends StatelessWidget {
-  final DateTime estimatedEventCausingDate;
-  final CreatedPillSheetValue? value;
-
-  const PillSheetModifiedHistoryCreatePillSheetAction({
-    Key? key,
-    required this.estimatedEventCausingDate,
-    required this.value,
-  }) : super(key: key);
-
-  Widget build(BuildContext context) {
-    return Container(
-      child: Padding(
-        padding: const EdgeInsets.only(top: 4, bottom: 4),
-        child: Row(
-          children: [
-            PillSheetModifiedHistoryDate(
-              estimatedEventCausingDate: estimatedEventCausingDate,
-              effectivePillNumber:
-                  PillSheetModifiedHistoryDateEffectivePillNumber
-                      .pillSheetCount(value?.pillSheetIDs ?? []),
-            ),
-            Spacer(),
-            Container(
-              width: PillSheetModifiedHistoryTakenActionLayoutWidths.trailing,
-              child: Row(
-                children: [
-                  Container(
-                    child: Text(
-                      "ピルシート追加",
-                      style: TextStyle(
-                        color: TextColor.main,
-                        fontSize: 12,
-                        fontFamily: FontFamily.japanese,
-                        fontWeight: FontWeight.w400,
-                      ),
-                      textAlign: TextAlign.start,
-                    ),
-                  ),
-                  Spacer(),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
